cloudformation_repos/requirements.txt
pyyaml==5.1
click==7.0
requests==2.21.0
terminaltables==3.1.0
colorclass==2.2.0